refactor: prover
