+++ crates/node_types/prover/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod prover;
 pub mod webserver;
+pub mod prover_engine;
+pub mod sequencer;
+pub mod syncer;
 
 pub use prover::{Config, Prover};
