+++ crates/node_types/prover/src/prover/tests.rs
@@ -1,6 +1,7 @@
 use super::*;
 use prism_common::test_transaction_builder::TestTransactionBuilder;
 use prism_keys::{CryptoAlgorithm, SigningKey, VerifyingKey};
+use prism_tree::proofs::Proof;
 use std::{self, sync::Arc, time::Duration};
 use tokio::spawn;
 
@@ -44,7 +45,7 @@ async fn test_posts_epoch_after_max_gap(algorithm: CryptoAlgorithm) {
         prover_handle.run().await.unwrap();
     });
 
-    let mut rx = prover.da.subscribe_to_heights();
+    let mut rx = prover.get_da().subscribe_to_heights();
 
     // Wait for initial blocks to be produced
     loop {
@@ -55,7 +56,7 @@ async fn test_posts_epoch_after_max_gap(algorithm: CryptoAlgorithm) {
     }
 
     // Ensure no gap proof has been created
-    assert!(prover.da.get_finalized_epoch(0).await.unwrap().is_none());
+    assert!(prover.get_da().get_finalized_epoch(0).await.unwrap().is_none());
 
     // Create and submit transactions
     let test_transactions = create_mock_transactions(algorithm, "test_service".to_string());
@@ -76,7 +77,7 @@ async fn test_posts_epoch_after_max_gap(algorithm: CryptoAlgorithm) {
             break;
         }
     }
-    let initial_epoch = prover.da.get_finalized_epoch(initial_epoch_height).await.unwrap().unwrap();
+    let initial_epoch = prover.get_da().get_finalized_epoch(initial_epoch_height).await.unwrap().unwrap();
 
     // Wait for gap length
     loop {
@@ -96,7 +97,7 @@ async fn test_posts_epoch_after_max_gap(algorithm: CryptoAlgorithm) {
         }
     }
     // Verify gap proof contents
-    let gap_proof = prover.da.get_finalized_epoch(current_epoch_height).await.unwrap().unwrap();
+    let gap_proof = prover.get_da().get_finalized_epoch(current_epoch_height).await.unwrap().unwrap();
     assert_eq!(
         gap_proof.height,
         initial_epoch.height + 1,
@@ -119,7 +120,8 @@ async fn test_validate_and_queue_update(algorithm: CryptoAlgorithm) {
 
     prover.clone().validate_and_queue_update(transaction.clone()).await.unwrap();
 
-    let pending_transactions = prover.pending_transactions.read().await;
+    let pending_tx_arc = prover.get_pending_transactions();
+    let pending_transactions = pending_tx_arc.read().await;
     assert_eq!(pending_transactions.len(), 2);
 }
 
@@ -222,7 +224,7 @@ async fn test_restart_sync_from_scratch(algorithm: CryptoAlgorithm) {
     let transactions = create_mock_transactions(algorithm, "test_service".to_string());
 
     for transaction in transactions {
-        prover.clone().validate_and_queue_update(transaction).await.unwrap();
+        prover.validate_and_queue_update(transaction).await.unwrap();
         while let Ok(new_block) = brx.recv().await {
             if new_block.epoch.is_some() {
                 break;
@@ -230,14 +232,14 @@ async fn test_restart_sync_from_scratch(algorithm: CryptoAlgorithm) {
         }
     }
 
-    assert_eq!(prover.clone().db.get_latest_epoch_height().unwrap(), 3);
+    assert_eq!(prover.get_db().get_latest_epoch_height().unwrap(), 3);
 
     let prover2 = Arc::new(Prover::new(db2.clone(), da_layer.clone(), &cfg).unwrap());
     let runner = prover2.clone();
     spawn(async move { runner.run().await.unwrap() });
 
     loop {
-        let epoch = prover2.clone().db.get_latest_epoch_height();
+        let epoch = prover2.get_db().get_latest_epoch_height();
         if epoch.is_ok() && epoch.unwrap() == 3 {
             assert_eq!(
                 prover.get_commitment().await.unwrap(),
@@ -264,7 +266,7 @@ async fn test_load_persisted_state(algorithm: CryptoAlgorithm) {
     let transactions = create_mock_transactions(algorithm, "test_service".to_string());
 
     for transaction in transactions {
-        prover.clone().validate_and_queue_update(transaction).await.unwrap();
+        prover.validate_and_queue_update(transaction).await.unwrap();
         while let Ok(new_block) = brx.recv().await {
             if new_block.epoch.is_some() {
                 break;
@@ -272,12 +274,12 @@ async fn test_load_persisted_state(algorithm: CryptoAlgorithm) {
         }
     }
 
-    assert_eq!(prover.clone().db.get_latest_epoch_height().unwrap(), 3);
+    assert_eq!(prover.get_db().get_latest_epoch_height().unwrap(), 3);
 
     let prover2 = Arc::new(Prover::new(db.clone(), da_layer.clone(), &cfg).unwrap());
     let runner = prover2.clone();
     spawn(async move { runner.run().await.unwrap() });
-    let epoch = prover2.clone().db.get_latest_epoch_height().unwrap();
+    let epoch = prover2.get_db().get_latest_epoch_height().unwrap();
     assert_eq!(epoch, 3);
     assert_eq!(
         prover.get_commitment().await.unwrap(),
