+++ crates/node_types/prover/src/sequencer.rs
@@ -1,0 +1,216 @@
+use anyhow::{Context, Result, bail};
+use jmt::KeyHash;
+use prism_common::{
+    account::Account,
+    digest::Digest,
+    operation::Operation,
+    transaction::Transaction,
+};
+use prism_da::{DataAvailabilityLayer, FinalizedEpoch};
+use prism_keys::SigningKey;
+use prism_storage::database::Database;
+use prism_tree::{
+    AccountResponse::*,
+    hasher::TreeHasher,
+    key_directory_tree::KeyDirectoryTree,
+    proofs::Proof,
+    snarkable_tree::SnarkableTree,
+};
+use std::sync::Arc;
+use tokio::sync::RwLock;
+
+use crate::prover_engine::ProverEngine;
+
+#[derive(Clone)]
+pub struct Sequencer {
+    db: Arc<Box<dyn Database>>,
+    da: Arc<dyn DataAvailabilityLayer>,
+    tree: Arc<RwLock<KeyDirectoryTree<Box<dyn Database>>>>,
+    pending_transactions: Arc<RwLock<Vec<Transaction>>>,
+    signing_key: SigningKey,
+    latest_epoch_da_height: Arc<RwLock<u64>>,
+    batcher_enabled: bool,
+}
+
+impl Sequencer {
+    pub fn new(
+        db: Arc<Box<dyn Database>>,
+        da: Arc<dyn DataAvailabilityLayer>,
+        signing_key: SigningKey,
+        latest_epoch_da_height: Arc<RwLock<u64>>,
+        batcher_enabled: bool,
+    ) -> Result<Self> {
+        let saved_epoch = match db.get_latest_epoch_height() {
+            Ok(height) => height + 1,
+            Err(_) => {
+                debug!("no existing epoch state found, starting at epoch 0");
+                0
+            }
+        };
+
+        let tree = Arc::new(RwLock::new(KeyDirectoryTree::load(
+            db.clone(),
+            saved_epoch,
+        )));
+
+        Ok(Sequencer {
+            db,
+            da,
+            tree,
+            pending_transactions: Arc::new(RwLock::new(Vec::new())),
+            signing_key,
+            latest_epoch_da_height,
+            batcher_enabled,
+        })
+    }
+
+    pub async fn run_batch_poster(&self) -> Result<()> {
+        let mut height_rx = self.da.subscribe_to_heights();
+
+        loop {
+            let height = height_rx.recv().await?;
+            trace!("received height {}", height);
+
+            let pending_transactions = {
+                let mut ops = self.pending_transactions.write().await;
+                std::mem::take(&mut *ops)
+            };
+
+            let tx_count = pending_transactions.len();
+
+            if !pending_transactions.clone().is_empty() {
+                match self.da.submit_transactions(pending_transactions).await {
+                    Ok(submitted_height) => {
+                        info!(
+                            "post_batch_loop: submitted {} transactions at height {}",
+                            tx_count, submitted_height
+                        );
+                    }
+                    Err(e) => {
+                        error!("post_batch_loop: Failed to submit transactions: {}", e);
+                    }
+                }
+            } else {
+                debug!(
+                    "post_batch_loop: No pending transactions to submit at height {}",
+                    height
+                );
+            }
+        }
+    }
+
+    pub async fn finalize_new_epoch(
+        &self,
+        epoch_height: u64,
+        transactions: Vec<Transaction>,
+        prover_engine: &Arc<ProverEngine>,
+    ) -> Result<u64> {
+        let mut tree = self.tree.write().await;
+        let batch = tree.process_batch(transactions)?;
+        batch.verify()?;
+
+        let (proof, compressed_proof) = prover_engine.prove_epoch(epoch_height, &batch, &self.db).await?;
+
+        let public_values = proof.public_values.to_vec();
+
+        let mut epoch_json = FinalizedEpoch {
+            height: epoch_height,
+            prev_commitment: batch.prev_root,
+            current_commitment: batch.new_root,
+            proof,
+            compressed_proof,
+            public_values,
+            signature: None,
+        };
+
+        epoch_json.insert_signature(&self.signing_key)?;
+
+        let da_height = self.da.submit_finalized_epoch(epoch_json.clone()).await?;
+        let mut latest_da_height = self.latest_epoch_da_height.write().await;
+        *latest_da_height = da_height;
+
+        self.db.add_epoch(&epoch_json)?;
+
+        info!("finalized new epoch at height {}", epoch_height);
+
+        Ok(da_height)
+    }
+
+    pub async fn execute_block(&self, transactions: Vec<Transaction>) -> Result<Vec<Proof>> {
+        debug!("executing block with {} transactions", transactions.len());
+
+        let mut proofs = Vec::new();
+
+        for transaction in transactions {
+            match self.process_transaction_internal(transaction.clone()).await {
+                Ok(proof) => proofs.push(proof),
+                Err(e) => {
+                    warn!(
+                        "Failed to process transaction: {:?}. Error: {}",
+                        transaction, e
+                    );
+                }
+            }
+        }
+
+        Ok(proofs)
+    }
+
+    async fn process_transaction_internal(&self, transaction: Transaction) -> Result<Proof> {
+        let mut tree = self.tree.write().await;
+        tree.process_transaction(transaction)
+    }
+
+    pub async fn validate_and_queue_update(&self, transaction: Transaction) -> Result<()> {
+        if !self.batcher_enabled {
+            bail!("Batcher is disabled, cannot queue transactions");
+        }
+
+        match transaction.operation {
+            Operation::RegisterService { .. } | Operation::CreateAccount { .. } => {
+                Account::default().process_transaction(&transaction)?;
+            }
+            Operation::AddKey { .. }
+            | Operation::RevokeKey { .. }
+            | Operation::AddData { .. }
+            | Operation::SetData { .. } => {
+                let account_response = self.get_account(&transaction.id).await?;
+
+                let Found(mut account, _) = account_response else {
+                    bail!("Account not found for id: {}", transaction.id)
+                };
+
+                account.process_transaction(&transaction)?;
+            }
+        };
+
+        let mut pending = self.pending_transactions.write().await;
+        pending.push(transaction);
+        Ok(())
+    }
+
+    pub async fn get_commitment(&self) -> Result<Digest> {
+        let tree = self.tree.read().await;
+        tree.get_commitment().context("Failed to get commitment")
+    }
+
+    pub async fn get_account(&self, id: &str) -> Result<prism_tree::AccountResponse> {
+        let tree = self.tree.read().await;
+        let key_hash = KeyHash::with::<TreeHasher>(id);
+
+        tree.get(key_hash)
+    }
+
+    pub fn get_pending_transactions(&self) -> Arc<RwLock<Vec<Transaction>>> {
+        self.pending_transactions.clone()
+    }
+    
+    pub fn get_db(&self) -> Arc<Box<dyn Database>> {
+        self.db.clone()
+    }
+    
+    pub async fn process_transaction(&self, transaction: Transaction) -> Result<Proof> {
+        let mut tree = self.tree.write().await;
+        tree.process_transaction(transaction)
+    }
+}
